requirements.txt
pillow~=11.2.1
scipy~=1.15.2